includes all view components

+++ src/views/MngUserPanel.java
@@ -262,18 +262,23 @@ public class MngUserPanel extends JPanel implements ActionListener, FocusListene
 			attemptUserCreation();
 		}
 		else if(source == deleteUsrBtn) {
-			String name = (String) userComboBox.getSelectedItem();
-			JFrame alert = new JFrame();
-			int response = JOptionPane.showConfirmDialog(alert, "This will delete the user " + name + " and all of their passwords. Do you wish to continue?","WARNING!!!", JOptionPane.YES_NO_OPTION, JOptionPane.WARNING_MESSAGE);
-			if(response == 0) {
-				boolean userDeleted = ctrl.deleteUser(name);
-				if (userDeleted == true) {
-					flashLbl.setText("User successfully deleted!");
-					flashLbl.setVisible(true);
-				}
-				else {
-					flashLbl.setText("ERROR: User could not be deleted");
-					flashLbl.setVisible(false);
+			int numUsers = conn.getRowCountFromTable("users");
+			if (numUsers < 2){
+				JOptionPane.showMessageDialog(null, "There must be at least one user in the database");
+			}
+			else {
+				String name = (String) userComboBox.getSelectedItem();
+				int response = JOptionPane.showConfirmDialog(null, "This will delete the user " + name + " and all of their passwords. Do you wish to continue?","WARNING!!!", JOptionPane.YES_NO_OPTION, JOptionPane.WARNING_MESSAGE);
+				if(response == 0) {
+					boolean userDeleted = ctrl.deleteUser(name);
+					if (userDeleted == true) {
+						flashLbl.setText("User successfully deleted!");
+						flashLbl.setVisible(true);
+					}
+					else {
+						flashLbl.setText("ERROR: User could not be deleted");
+						flashLbl.setVisible(false);
+					}
 				}
 			}
 		}
